Extensive updates and changes

+++ middleware/require_auth.go
@@ -56,7 +56,7 @@ func RequireAuth(next http.Handler) http.Handler {
 			return
 		}
 
-		ctx := context.WithValue(r.Context(), userIDKey, uint(userId))
+		ctx := context.WithValue(r.Context(), userIDKey, uint64(userId))
 		r = r.WithContext(ctx)
 
 		next.ServeHTTP(w, r)
