+++ services/setting.go
@@ -19,7 +19,7 @@ func NewSettingService(db *gorm.DB, tbClient tb.Client) *SettingService {
 	}
 }
 
-func (s *SettingService) GetAll(userID uint) ([]models.Setting, error) {
+func (s *SettingService) GetAll(userID uint64) ([]models.Setting, error) {
 	var settings []models.Setting
 	if err := s.db.Where("user_id = ?", userID).Or("type = ?", "app").Find(&settings).Error; err != nil {
 		return nil, err
@@ -27,12 +27,12 @@ func (s *SettingService) GetAll(userID uint) ([]models.Setting, error) {
 	return settings, nil
 }
 
-func (s *SettingService) SetSetting(userID uint, setting string, value string) error {
+func (s *SettingService) SetSetting(userID uint64, setting string, value string) error {
 	s.LogActivity(models.ActivityTypeAdmin, fmt.Sprintf("update setting %s to %s", setting, value), userID)
 	return s.db.Exec("INSERT INTO settings (user_id, setting, value) VALUES (?, ?, ?) ON DUPLICATE KEY UPDATE value = ?", setting, value, value).Error
 }
 
-func (s *SettingService) GetSetting(userID uint, setting string) (models.Setting, error) {
+func (s *SettingService) GetSetting(userID uint64, setting string) (models.Setting, error) {
 	var dbSetting models.Setting
 	err := s.db.Where("user_id", userID).Where("setting = ?", setting).First(dbSetting).Error
